requirements.txt
Flask==3.1.2
boto3==1.40.55
PyGithub==2.6.0
WTForms==3.2.1
Jinja2==3.1.5
PyYAML==6.0.2
gunicorn==23.0.0
python-dotenv==1.0.1
email-validator==2.2.0
